requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
constantly==15.1.0
coverage==6.3.2
cryptography==36.0.2
daphne==3.0.2
Django==4.0.3
django-filter==21.1
django-nextjs==2.1.1
djangorestframework==3.13.1
Faker==12.0.1
frozenlist==1.3.0
hyperlink==21.0.0
hypothesis==6.39.4
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
iniconfig==1.1.1
Markdown==3.3.6
mixer==7.2.2
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
websockets==10.2
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
